pyproject.toml
[project]
name = "shadow-agentic"
version = "0.1.0"
description = ""
authors = [
    {name = "chadxeth",email = "junmtan.eth@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain-openai (>=0.3.3,<0.4.0)",
    "langgraph (>=0.2.69,<0.3.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.17,<0.4.0)",
    "pika (>=1.3.2,<2.0.0)",
    "qdrant-client (>=1.13.2,<2.0.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "jsonpath-ng (>=1.7.0,<2.0.0)"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "shadow_agentic", from = "." }
]

